Rename string to str to align it with number and bool modules
