+++ lib/src/metta/runner/environment.rs
@@ -455,8 +455,8 @@ fn git_catalog_from_cfg_atom(atom: &ExpressionAtom, env: &Environment) -> Result
     let refresh_time = refresh_time.ok_or_else(|| format!("Error in environment.metta. \"refreshTime\" property required for #gitCatalog"))?
         .parse::<u64>().map_err(|e| format!("Error in environment.metta.  Error parsing \"refreshTime\": {e}"))?;
 
-    let catalog_name = crate::metta::runner::string::strip_quotes(catalog_name);
-    let catalog_url = crate::metta::runner::string::strip_quotes(catalog_url);
+    let catalog_name = crate::metta::runner::str::strip_quotes(catalog_name);
+    let catalog_url = crate::metta::runner::str::strip_quotes(catalog_url);
 
     let mut managed_remote_catalog = LocalCatalog::new(caches_dir, catalog_name).unwrap();
     let remote_catalog = GitCatalog::new(caches_dir, env.fs_mod_formats.clone(), catalog_name, catalog_url, refresh_time).unwrap();
@@ -474,7 +474,7 @@ fn include_path_from_cfg_atom(atom: &ExpressionAtom, env: &Environment) -> Resul
         None => return Err(format!("Error in environment.metta. #includePath missing path value"))
     };
     let path = <&crate::SymbolAtom>::try_from(path_atom)?.name();
-    let path = crate::metta::runner::string::strip_quotes(path);
+    let path = crate::metta::runner::str::strip_quotes(path);
 
     //TODO-FUTURE: In the future we may want to replace dyn-fmt with strfmt, and do something a
     // little bit nicer than this
