+++ lib/src/metta/runner/stdlib/string.rs
@@ -1,7 +1,7 @@
 use crate::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
-use crate::metta::runner::string::*;
+use crate::metta::runner::str::*;
 use super::{grounded_op, atom_to_string, unit_result, regex};
 
 use std::convert::TryInto;
