+++ lib/src/metta/runner/str.rs
